requirement.txt
Django==5.0.3
sqlparse==0.4.4
gunicorn==21.2.0
requests==2.31.0
nested_lookup==0.2.25
redis==5.0.2
python-decouple==3.8
psycopg2==2.9.9
python-dateutil==2.9.0
xlsxwriter==3.2.0
django-cors-headers==4.3.1
whitenoise==6.6.0
reportlab==4.1.0
pytz==2024.1
djangorestframework==3.15.1
django-colorfield==0.11.0







